Move StepContext to top level
