+++ runners/core-java/src/main/java/org/apache/beam/runners/core/StepContext.java
@@ -15,56 +15,56 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package org.apache.beam.fn.harness.fake;
+package org.apache.beam.runners.core;
 
 import java.io.IOException;
-import org.apache.beam.runners.core.StepContext;
-import org.apache.beam.runners.core.StateInternals;
-import org.apache.beam.runners.core.TimerInternals;
 import org.apache.beam.sdk.coders.Coder;
+import org.apache.beam.sdk.transforms.DoFn.WindowedContext;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
 import org.apache.beam.sdk.util.WindowedValue;
 import org.apache.beam.sdk.values.TupleTag;
 
 /**
- * A fake {@link StepContext} factory that performs no-ops.
+ * Per-step, per-key context used for retrieving state.
  */
-public class FakeStepContext implements StepContext {
-  @Override
-  public String getStepName() {
-    return "TODO";
-  }
+public interface StepContext {
 
-  @Override
-  public String getTransformName() {
-    return "TODO";
-  }
+  /**
+   * The name of the step.
+   */
+  String getStepName();
 
-  @Override
-  public void noteOutput(WindowedValue<?> output) {
-  }
+  /**
+   * The name of the transform for the step.
+   */
+  String getTransformName();
 
-  @Override
-  public void noteOutput(TupleTag<?> tag, WindowedValue<?> output) {
-  }
+  /**
+   * Hook for subclasses to implement that will be called whenever
+   * {@link WindowedContext#output}
+   * is called.
+   */
+  void noteOutput(WindowedValue<?> output);
 
-  @Override
-  public <T, W extends BoundedWindow> void writePCollectionViewData(
+  /**
+   * Hook for subclasses to implement that will be called whenever
+   * {@link WindowedContext#output}
+   * is called.
+   */
+  void noteOutput(TupleTag<?> tag, WindowedValue<?> output);
+
+  /**
+   * Writes the given {@code PCollectionView} data to a globally accessible location.
+   */
+  <T, W extends BoundedWindow> void writePCollectionViewData(
       TupleTag<?> tag,
       Iterable<WindowedValue<T>> data,
       Coder<Iterable<WindowedValue<T>>> dataCoder,
       W window,
-      Coder<W> windowCoder) throws IOException {
-  }
+      Coder<W> windowCoder)
+          throws IOException;
 
-  @Override
-  public StateInternals stateInternals() {
-    throw new UnsupportedOperationException();
-  }
+  StateInternals stateInternals();
 
-  @Override
-  public TimerInternals timerInternals() {
-    throw new UnsupportedOperationException();
-  }
+  TimerInternals timerInternals();
 }
